# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: ADHD Productivity System API
  description: |
    A comprehensive API for managing tasks, focus sessions, and productivity metrics specifically designed for individuals with ADHD.
    
    ## Key Features
    - **ADHD-Centered Design**: All endpoints are optimized for users with ADHD
    - **Focus Session Management**: Pomodoro-style focus sessions with ADHD-friendly customizations
    - **Task Decomposition**: Smart task breaking for executive function support
    - **Real-time Updates**: WebSocket support for live session synchronization
    - **Analytics & Insights**: Productivity tracking with ADHD-specific metrics
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    - Standard endpoints: 100 requests/minute per user
    - Authentication endpoints: 5 requests/minute per IP
    - File uploads: 10 requests/minute per user
    
    ## ADHD-Specific Considerations
    - All response times are optimized for attention management
    - Error messages are clear and actionable
    - Minimal cognitive load in API design
    - Support for executive function assistance
  version: 1.0.0
  contact:
    name: ADHD Productivity System Team
    email: api-support@adhd-productivity.dev
    url: https://docs.adhd-productivity.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Create a new user account with ADHD-specific profile options.
        
        **ADHD Considerations:**
        - Optional ADHD type classification for personalized experience
        - Cognitive load preferences for UI/UX customization
        - Timezone support for time management features
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
            examples:
              basic_registration:
                summary: Basic user registration
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
                  firstName: "Jane"
                  lastName: "Doe"
                  timezone: "America/New_York"
              adhd_specific:
                summary: Registration with ADHD profile
                value:
                  email: "adhd.user@example.com"
                  password: "SecurePass123!"
                  firstName: "Alex"
                  lastName: "Smith"
                  timezone: "Europe/London"
                  adhdType: "Combined"
                  theme: "Dark"
                  cognitiveLoadPreference: "Low"
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive access token
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Task Management Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRegistration:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: Strong password with mixed characters
          example: "SecurePass123!"
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          description: User's first name
          example: "Jane"
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          description: User's last name
          example: "Doe"
        timezone:
          type: string
          description: User's timezone (IANA format)
          example: "America/New_York"
        adhdType:
          type: string
          enum: [Inattentive, Hyperactive, Combined, NotSpecified]
          description: ADHD type for personalization
          example: "Combined"
        theme:
          type: string
          enum: [Light, Dark, Auto, HighContrast]
          description: UI theme preference
          example: "Dark"
        cognitiveLoadPreference:
          type: string
          enum: [Low, Medium, High]
          description: Preferred cognitive load level
          example: "Low"
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"
        rememberMe:
          type: boolean
          default: false
          description: Extended session duration
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            refreshToken:
              type: string
              description: Refresh token for renewing access
              example: "def50200..."
            expiresIn:
              type: integer
              description: Token expiration time in seconds
              example: 3600
        message:
          type: string
          example: "Authentication successful"
        timestamp:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"

    # User Schema
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "Jane"
        lastName:
          type: string
          example: "Doe"
        timezone:
          type: string
          example: "America/New_York"
        adhdType:
          type: string
          enum: [Inattentive, Hyperactive, Combined, NotSpecified]
          example: "Combined"
        theme:
          type: string
          enum: [Light, Dark, Auto, HighContrast]
          example: "Dark"
        createdAt:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
        lastLoginAt:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"

    # Task Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        message:
          type: string
          description: Human-readable error message
          example: "Invalid input data"
        error:
          type: string
          description: Technical error details
          example: "Validation failed for field 'email'"
        timestamp:
          type: string
          format: date-time
          example: "2024-12-22T10:00:00Z"
        validationErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Must be a valid email address"
              code:
                type: string
                example: "INVALID_FORMAT"
        errorCode:
          type: string
          description: Application-specific error code
          example: "VALIDATION_ERROR"
        requestId:
          type: string
          description: Request tracking ID for support
          example: "req_123456789"
...